fix --no-default-features build
